feat(drawer): optimize appendtobody by use built-in teleport
